Extract a single card as separate component
