+++ frontend/src/app/components/wp-card-view/styles/wp-card-view-vertical.sass
@@ -10,9 +10,3 @@
   // independently of whether we scroll or not.
   overflow-y: scroll
   @include styled-scroll-bar
-
-  .wp-card
-    margin-top: 10px
-    // Take care that the shadow of the last element is still visible
-    &:last-of-type
-      margin-bottom: 3px
