+++ frontend/src/app/components/wp-card-view/styles/wp-card-view.component.sass
@@ -1,86 +1,7 @@
-@import 'helpers'
-
-.wp-card
-  display: flex
-  flex-direction: column
-  user-select: none
-  width: 100%
-  border: 1px solid var(--widget-box-block-border-color)
-  border-radius: 2px
-  padding: 10px
-  position: relative
-  box-shadow: 1px 1px 3px 0px lightgrey
-  background: var(--body-background)
-  font-size: var(--card-font-size)
-  max-width: 400px
-
-  &:hover
-    box-shadow: 0px 0px 10px lightgrey
-
-  &.-new
-    padding-right: 25px
-
-  &.-checked
-    background-color: var(--table-row-highlighting-color)
-
-.wp-card--content:not(.-new)
-  display: grid
-  grid-template-columns: auto 1fr auto
-  grid-template-rows: auto 1fr auto
-  grid-row-gap: 10px
-  grid-template-areas: "type type type" "subject subject subject" "attributeTag avatar idlink"
-  overflow: hidden
-  flex-grow: 1
-
-  .wp-card--type
-    grid-area: type
-  .wp-card--subject
-    grid-area: subject
-    max-width: 230px
-    @include text-shortener
-    white-space: normal
-  .wp-card--assignee
-    grid-area: avatar
-    place-self: center left
-  .wp-card--id
-    grid-area: idlink
-    place-self: center right
-  .wp-card--status
-    grid-area: attributeTag
-    max-width: 120px
-    margin-right: 5px
-    overflow: hidden
-
-.wp-card--highlighting
-  width: 5px
-  height: 100%
-  position: absolute
-  top: 0
-  left: 0
-  border-radius: 2px 0 0 2px
-
 .wp-inline-create-button
   font-size: 0.9rem
   padding-top: 1rem
   text-align: center
 
-.wp-card--inline-buttons
-  position: absolute
-  right: 0
-  top: 5px
-  opacity: 0
-
-  &.-show, .wp-card:hover &
-    opacity: 1
-
 .wp-inline-create--reference-container
   margin-bottom: 3rem
-
-.wp-card--cover-image
-  display: block
-  margin: -10px -10px 10px
-  width: calc(100% + 10px + 10px)
-  max-width: calc(100% + 10px + 10px)
-
-  flex-basis: 200px
-  object-fit: cover
