+++ frontend/src/app/components/wp-card-view/wp-card-view.component.html
@@ -8,76 +8,16 @@
     </ndc-dynamic>
   </div>
 
-  <div class="wp-card"
-       *ngFor="let wp of workPackages; trackBy:trackByHref"
-       [attr.data-is-new]="wp.isNew || undefined"
-       [attr.data-work-package-id]="wp.id"
-       [attr.data-class-identifier]="classIdentifier(wp)"
-       [ngClass]="cardClasses(wp)">
-
-    <div class="wp-card--highlighting"
-        [ngClass]="cardHighlightingClass(wp)">
-    </div>
-
-    <div class="wp-card--inline-buttons">
-      <a class="wp-card--inline-cancel-button -no-decoration"
-         *ngIf="wp.isNew || cardsRemovable"
-         [ngClass]="{ '-show': wp.isNew }"
-         [title]="text.removeCard"
-         (accessibleClick)="removeCard(wp)">
-        <op-icon icon-classes="icon icon-close"></op-icon>
-      </a>
-
-      <a class="-no-decoration"
-         *ngIf="!wp.isNew && showInfoButton"
-         [title]="text.detailsView"
-         (accessibleClick)="openSplitScreen(wp)">
-        <op-icon icon-classes="icon icon-info2"></op-icon>
-      </a>
-    </div>
-
-    <edit-form [resource]="wp"
-                         [inEditMode]="wp.isNew"
-                         *ngIf="wp.isNew">
-      <div class="wp-card--content -new">
-        <editable-attribute-field [resource]="wp"
-                       [wrapperClasses]="'work-packages--type-selector'"
-                       [fieldName]="'type'"
-                       class="wp-card--type">
-        </editable-attribute-field>
-        <editable-attribute-field [resource]="wp"
-                       fieldName="subject"
-                       class="wp-card--subject -bold">
-        </editable-attribute-field>
-      </div>
-    </edit-form>
-
-    <img *ngIf="this.bcfSnapshotPath(wp)" [src]="this.bcfSnapshotPath(wp)" class="wp-card--cover-image">
-
-    <div *ngIf="!wp.isNew"
-         class="wp-card--content">
-      <span [textContent]="wpTypeAttribute(wp)"
-          class="wp-card--type"
-          [ngClass]="typeHighlightingClass(wp)"></span>
-      <a uiSref="work-packages.show"
-         [uiParams]="{workPackageId: wp.id}"
-         class="wp-card--id"
-         [ngClass]="uiStateLinkClass">
-        #{{wp.id}}
-      </a>
-      <span [textContent]="wpSubject(wp)"
-            class="wp-card--subject"></span>
-      <wp-status-button *ngIf="showStatusButton"
-                        [workPackage]="wp"
-                        class="wp-card--status">
-      </wp-status-button>
-      <user-avatar *ngIf="wp.assignee"
-                   [user]="wp.assignee"
-                   data-class-list="avatar-mini"
-                   class="wp-card--assignee">
-      </user-avatar>
-    </div>
-  </div>
+  <wp-single-card *ngFor="let wp of workPackages; trackBy:trackByHref"
+                  [workPackage]="wp"
+                  [showInfoButton]="showInfoButton"
+                  [showStatusButton]="showStatusButton"
+                  [showRemoveButton]="cardsRemovable"
+                  [highlightingMode]="highlightingMode"
+                  [draggable]="this.canDragOutOf(wp)"
+                  [orientation]="orientation"
+                  (onRemove)="removeCard(wp)">
+  </wp-single-card>
 </div>
 
 <div *ngIf="showEmptyResultsBox && isResultEmpty">
