+++ frontend/src/app/components/wp-card-view/wp-card-view.component.ts
@@ -5,7 +5,6 @@ import {
   Component,
   ElementRef,
   EventEmitter,
-  Inject,
   Injector,
   Input,
   OnInit,
@@ -21,8 +20,6 @@ import {I18nService} from "core-app/modules/common/i18n/i18n.service";
 import {WorkPackageInlineCreateService} from "core-components/wp-inline-create/wp-inline-create.service";
 import {WorkPackageCreateService} from "core-components/wp-new/wp-create.service";
 import {AngularTrackingHelpers} from "core-components/angular/tracking-functions";
-import {HalResourceNotificationService} from "core-app/modules/hal/services/hal-resource-notification.service";
-import {Highlighting} from "core-components/wp-fast-table/builders/highlighting/highlighting.functions";
 import {CardHighlightingMode} from "core-components/wp-fast-table/builders/highlighting/highlighting-mode.const";
 import {AuthorisationService} from "core-app/modules/common/model-auth/model-auth.service";
 import {StateService} from "@uirouter/core";
@@ -35,7 +32,6 @@ import {WorkPackageViewSelectionService} from "core-app/modules/work_packages/ro
 import {CardViewHandlerRegistry} from "core-components/wp-card-view/event-handler/card-view-handler-registry";
 import {WorkPackageCardViewService} from "core-components/wp-card-view/services/wp-card-view.service";
 import {WorkPackageCardDragAndDropService} from "core-components/wp-card-view/services/wp-card-drag-and-drop.service";
-import {checkedClassName, uiStateLinkClass} from "core-components/wp-fast-table/builders/ui-state-link-builder";
 import {WorkPackageNotificationService} from "core-app/modules/work_packages/notifications/work-package-notification.service";
 
 export type CardViewOrientation = 'horizontal'|'vertical';
@@ -75,11 +71,7 @@ export class WorkPackageCardViewComponent  implements OnInit, AfterViewInit {
       title: this.I18n.t('js.work_packages.no_results.title'),
       description: this.I18n.t('js.work_packages.no_results.description')
     },
-    detailsView: this.I18n.t('js.button_open_details')
   };
-
-  public uiStateLinkClass:string = uiStateLinkClass;
-  public checkedClassName:string = checkedClassName;
 
   /** Inline create / reference properties */
   public canAdd = false;
@@ -160,68 +152,6 @@ export class WorkPackageCardViewComponent  implements OnInit, AfterViewInit {
 
   ngOnDestroy():void {
       this.cardDragDrop.destroy();
-  }
-
-  public openSplitScreen(wp:WorkPackageResource) {
-    let classIdentifier = this.classIdentifier(wp);
-    this.wpTableSelection.setSelection(wp.id!, this.cardView.findRenderedCard(classIdentifier));
-    this.$state.go(
-      'work-packages.list.details',
-      {workPackageId: wp.id!}
-    );
-  }
-
-  public wpTypeAttribute(wp:WorkPackageResource) {
-    return wp.type.name;
-  }
-
-  public wpSubject(wp:WorkPackageResource) {
-    return wp.subject;
-  }
-
-  public isSelected(wp:WorkPackageResource):boolean {
-    return this.wpTableSelection.isSelected(wp.id!);
-  }
-
-  public classIdentifier(wp:WorkPackageResource) {
-    return this.cardView.classIdentifier(wp);
-  }
-
-  public bcfSnapshotPath(wp:WorkPackageResource) {
-    let vp = _.get(wp, 'bcf.viewpoints[0]');
-    if (vp) {
-      return this.pathHelper.attachmentDownloadPath(vp.id, vp.file_name);
-    } else {
-      return null;
-    }
-  }
-
-
-  public cardClasses(wp:WorkPackageResource) {
-    let classes = this.isSelected(wp) ? checkedClassName : '';
-    classes += this.canDragOutOf(wp) ? ' -draggable' : '';
-    classes += wp.isNew ? ' -new' : '';
-    classes += ' wp-card-' + wp.id;
-    return classes;
-  }
-
-  public cardHighlightingClass(wp:WorkPackageResource) {
-    return this.cardHighlighting(wp);
-  }
-
-  public typeHighlightingClass(wp:WorkPackageResource) {
-    return this.attributeHighlighting('type', wp);
-  }
-
-  private cardHighlighting(wp:WorkPackageResource) {
-    if (['status', 'priority', 'type'].includes(this.highlightingMode)) {
-      return Highlighting.backgroundClass(this.highlightingMode, wp[this.highlightingMode].id);
-    }
-    return '';
-  }
-
-  private attributeHighlighting(type:string, wp:WorkPackageResource) {
-    return Highlighting.inlineClass(type, wp.type.id!);
   }
 
   public get workPackages():WorkPackageResource[] {
